Finished Brick Environment
I haven't added the music to the scene yet, but it's downloaded into the SFX folder as well as a goal sound. There still needs to be a script for the goal and for scoring. Once those are created I will be able to finish the UI. I was also having trouble with the particle system on the goal showing when you push play.