Finalizing everything
Removed a few test scenes and combined everything into a nice folder structure which should allow for us to cleanly turn this in. I am not sure what assets you need for the main menu. additionally we need to connect the main menu with the game scene so it starts the game.